--- capture-sdk/sdk/src/main/java/net/gini/android/capture/ClickListenerExt.kt
@@ -1,7 +1,0 @@
-package net.gini.android.capture
-
-import android.view.View
-
-fun View.onClick(click: (View) -> Unit) {
-    setOnClickListener(IntervalClickListener(click))
-}
