+++ capture-sdk/sdk/src/main/java/net/gini/android/capture/internal/ui/IntervalClickListener.kt
@@ -1,0 +1,23 @@
+package net.gini.android.capture.internal.ui
+
+import android.view.View
+
+
+class IntervalClickListener(
+    private val click: View.OnClickListener
+) : View.OnClickListener {
+
+    private val interval = 500L
+
+    override fun onClick(view: View) {
+        if (isEnabled) {
+            isEnabled = false
+            view.postDelayed(enable, interval)
+            click.onClick(view)
+        }
+    }
+
+    var isEnabled = true
+    private val enable =
+        Runnable { isEnabled = true }
+}
